+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # Changelog
+
+ - Fichiers handles et index créés et init.